==> Drug_Repositioning/8_new_training/PREDICT_data/1_gold_indep_leraning_asso.ipynb <==
import pandas as pd
import numpy as numpy
from itertools import product
from datetime import datetime

asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/2_gold_asso.tsv")
drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/4_gold_se_by_indep.tsv")
disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_disease_feature.tsv")
indep = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/3_indep_asso.tsv")

print len(asso)
print len(indep)

gold_dr = asso['DrugBank_ID'].drop_duplicates()
gold_di = asso['OMIM_ID'].drop_duplicates()
indep_dr = indep['drug'].drop_duplicates()
indep_di = indep['disease'].drop_duplicates()

print len(gold_dr)
print len(gold_di)
print len(indep_dr)
print len(indep_di)

# gold_drug, gold_disease에서 나올 수 있는 모든 association pair
all_pair = pd.DataFrame(list(product(gold_dr, gold_di)), columns = ['drug_id', 'disease_id'])

# indep_drug, indep_disease에서 나올 수 있는 모든 association pair
indep_all_pair = pd.DataFrame(list(product(indep_dr, indep_di)), columns = ['drug_id', 'disease_id'])
print len(indep_all_pair)
# 일단 모든 association을 0으로 표시한다.
indep_all_pair['association'] = 0
indep_all_pair.head(1)

# gold asso(positive)에 포함되어 있는 association을 independent에서 제거 (아마 indep의 negative에서 제거 될 거임.)
for index, row in indep_all_pair.iterrows():
    dr_id = row['drug_id']
    di_id = row['disease_id']
    found = asso[(asso.DrugBank_ID == dr_id) & (asso.OMIM_ID == di_id)]
    if found.empty:
        continue
    else:
        indep_all_pair.loc[index, 'association'] = 2 # gold의 positive와 겹치면 association을 2로 set한다.

print len(indep_all_pair)
print len(indep_all_pair[(indep_all_pair.association ==2)]) # gold positive와 겹치는 independent의 association 수
indep_pair = indep_all_pair[~(indep_all_pair.association ==2)] # gold positive와 겹치는 independent의 association을 제거함.
print len(indep_pair)

# positive인 independent 정보를 all pair(gold positive와는 중복 제거된)에 표시한다.
for index, row in indep.iterrows():
    dr = row['drug']
    di = row['disease']
    found = indep_pair[(indep_pair.drug_id == dr) & (indep_pair.disease_id == di)]
    indep_pair.loc[found.index, 'association'] = 1

print len(indep_pair)
print len(indep_pair[indep_pair.association ==1])
print len(indep_pair[indep_pair.association ==0])
print len(indep_pair.drop_duplicates('drug_id'))
print len(indep_pair.drop_duplicates('disease_id'))

temp = indep_pair[indep_pair['drug_id'].isin(gold_dr)] # independent data중 gold dr과 연관이 있는 association
print len(temp.drop_duplicates('drug_id'))
print len(temp.drop_duplicates('disease_id'))
print len(temp[temp.association ==1]) # 그 중 positive association의 수

p = indep_pair[indep_pair.association ==1]
n= indep_pair[indep_pair.association ==0]
print len(n[n['drug_id'].isin(gold_dr)])
print len(n[n['drug_id'].isin(gold_dr)].drop_duplicates('drug_id'))
print len(n[~n['drug_id'].isin(gold_dr)])
print len(n[~n['drug_id'].isin(gold_dr)].drop_duplicates('drug_id'))

indep_pair.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/3_2_indep_learning_asso.tsv", sep='\t', index=False)
indep_pair.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/PREDICT_data/3_2_indep_learning_asso.tsv", sep='\t', index=False)

print len(all_pair)
print len(indep_pair)

# 2017.4.19. 새로운 방법으로 negative check
all_pair.set_index(['drug_id', 'disease_id'], inplace=True)
indep_assos = [tuple(x) for x in indep_pair[['drug_id', 'disease_id']].values]
a = all_pair[~all_pair.index.isin(indep_assos)]  # indep association이 포함되지 않은 gold association
print len(a)

pos_assos = [tuple(x) for x in asso[['DrugBank_ID', 'OMIM_ID']].values] # positive association
print len(pos_assos)
pos_asso = a.filter(items = pos_assos, axis = 0) # positive association
neg_asso = a[~a.index.isin(a.index.isin(pos_assos))] # negative association
print len(pos_asso)
print len(neg_asso)

# gold association의 negativie 구함.(positive에도 없고, independent에도 없는 association을 찾는것.)
neg = pd.DataFrame()
print len(all_pair)
cnt = 0
print "start : " + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
for index, row in all_pair.iterrows():
    dr_id = row['drug_id']
    di_id = row['disease_id']
    found = asso[(asso.DrugBank_ID == dr_id) & (asso.OMIM_ID == di_id)] # asso는 gold positive
    found2 = indep_pair[(indep_pair.drug_id == dr_id) & (indep_pair.disease_id == di_id)] # indep의 모든 association
    if found.empty & found2.empty:
        neg = neg.append(row)
    cnt = cnt + 1
    if cnt % 10000 ==0:
        print "cnt : {}  ".format(cnt) + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print "end : " + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print len(neg)

asso = asso.rename(columns={'DrugBank_ID':'drug_id', 'OMIM_ID':'disease_id'})

asso.head(2)

neg.head(2)

int_di = list()
for i in neg['disease_id']:
    int_di.append(int(i))

neg['disease_id'] = int_di
neg.head(2)

asso['association'] = 1
neg['association'] = 0

learning_asso = pd.concat([asso, neg])
print len(learning_asso)
print len(learning_asso.drop_duplicates())
print len(learning_asso.drop_duplicates('drug_id'))
print len(learning_asso.drop_duplicates('disease_id'))
print len(learning_asso[learning_asso.association ==1])
print len(learning_asso[learning_asso.association ==0])

learning_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/PREDICT_data/2_2_gold_learning_asso.tsv", sep='\t', index=False)
learning_asso.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/2_2_gold_learning_asso.tsv", sep='\t', index=False)

##### descriptor 구축(하다맘 - 이 방법 말고 다른 파일 참조)

learning_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/2_2_gold_learning_asso.tsv")

print len(learning_asso)
print len(learning_asso[learning_asso.association ==1])
print len(learning_asso[learning_asso.association ==0])

learning_desc = pd.DataFrame()
start = 0
end = 20000
cnt = start
print len(learning_asso)
print "start : " + datetime.now().strftime('%Y-%m-%d %H:%M:%S')    
for index, row in learning_asso[start:end].iterrows():
    dr_id = row['drug_id']
    di_id = row['disease_id']
    found = o_desc[(o_desc.drug_id == dr_id) & (o_desc.disease_id == di_id)]
    learning_desc = learning_desc.append(found)
    cnt = cnt +1
    if cnt % 10000 ==0:
        print "cnt : {}  ".format(cnt) + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print "end : " + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
learning_desc.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/1_gold_descriptor_{}_{}.tsv".format(start, end-1),
                     sep='\t', index=False)

print len(learning_desc)
print len(desc2)
print len(desc3)
print len(desc4)
print len(desc5)
print len(desc6)

desc1 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/1_gold_descriptor_0_19999.tsv")
desc2 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/1_gold_descriptor_20000_39999.tsv")
desc3 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/1_gold_descriptor_40000_59999.tsv")
desc4 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/1_gold_descriptor_60000_79999.tsv")
desc5 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/1_gold_descriptor_80000_99999.tsv")
desc6 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/1_gold_descriptor_100000_103145.tsv")'''

all_desc = pd.concat([desc1, desc2, desc3, desc4, desc5, desc6])

print len(all_desc)
print len(all_desc.drop_duplicates())
print len(all_desc[all_desc.association ==1])
print len(all_desc[all_desc.association == 0])

print len(all_desc['drug_id'].drop_duplicates())
print len(all_desc['disease_id'].drop_duplicates())

print len(learning_asso['drug_id'].drop_duplicates())

learning_asso[~learning_asso['drug_id'].isin(all_desc['drug_id'])].drop_duplicates('drug_id')

o_desc[o_desc.drug_id == 'DB09026']

print len(learning_asso[learning_asso.drug_id == 'DB09026'])
print len(learning_asso[learning_asso.drug_id == 'DB04272'])
print len(learning_asso[learning_asso.drug_id == 'DB01378'])
print len(learning_asso[learning_asso.drug_id == 'DB00114'])
print len(learning_asso[learning_asso.drug_id == 'DB00533'])
print len(learning_asso[learning_asso.drug_id == 'DB03904'])
print len(learning_asso[learning_asso.drug_id == 'DB00580'])
