==> Drug_Repositioning/8_new_training/PREDICT_data/2_binary_matrix_gold_indep_drug_original_indi.ipynb <==
import pandas as pd
import numpy as np
from itertools import product
from datetime import datetime

gold_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/2_2_gold_learning_asso.tsv")
indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/3_2_indep_learning_asso.tsv")

print len(gold_asso)
print len(gold_asso[gold_asso.association ==1])
print len(gold_asso[gold_asso.association ==0])

drug_se = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/4_gold_se_by_indep.tsv")
indep_se = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/3_indep_drug_se.tsv")
di_ge = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col = 0)

drug_se.head(2)

print len(drug_se['umls_id'].drop_duplicates())
print len(indep_se['umls_id'].drop_duplicates())

drug-se binary matrix 만들기

umls_ids = drug_se['umls_id'].sort_values().drop_duplicates()
gold_drug = drug_se['DrugBank_ID'].drop_duplicates()
indep_drug = indep_se['DrugBank_ID'].drop_duplicates()

binary_drug = pd.DataFrame(columns=gold_drug.tolist(), index=umls_ids.tolist())
binary_drug = binary_drug.fillna(0)
print len(binary_drug)
binary_drug.head(1)

f = lambda x: 1 if x else 0
    
for dg_id in gold_drug:
    dg_df = drug_se[drug_se.DrugBank_ID == dg_id]
    umls_id = dg_df['umls_id']
    flags = umls_ids.isin(umls_id)
    oz_flags = map(f, flags)
    binary_drug[dg_id] = oz_flags

binary_drug.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/4_2_gold_se_by_indep_matrix.tsv", sep='\t')
binary_drug.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/4_2_gold_se_by_indep_matrix.tsv", sep='\t')

indep drug-se binary matrix 만들기

binary_indep = pd.DataFrame(columns=indep_drug.tolist(), index = umls_ids.tolist())
binary_indep = binary_indep.fillna(0)
print len(binary_indep)
binary_indep.head(1)

f = lambda x: 1 if x else 0
    
for dg_id in indep_drug:
    dg_df = indep_se[indep_se.DrugBank_ID == dg_id]
    umls_id = dg_df['umls_id']
    flags = umls_ids.isin(umls_id)
    oz_flags = map(f, flags)
    binary_indep[dg_id] = oz_flags

binary_indep.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/3_indep_drug_se_matrix.tsv", sep='\t')
binary_indep.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/3_indep_drug_se_matrix.tsv", sep='\t')

drug-indication binary matrix 만들기

gold_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/6_3_gold_learning_asso_indi.tsv")
indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/7_2_indep_learning_asso_indi.tsv")

print len(gold_asso)
print len(gold_asso[gold_asso.association ==1])
print len(gold_asso[gold_asso.association ==0])

drug_indi = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/sider_indication/3_3_gold_indication_indep_ver.tsv")
indep_indi = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/sider_indication/4_indep_indication.tsv")

drug_indi.head(2)

print len(drug_indi['UMLS_concept_id_for_MedDRA'].drop_duplicates())
print len(indep_indi['UMLS_concept_id_for_MedDRA'].drop_duplicates())

umls_ids = drug_indi['UMLS_concept_id_for_MedDRA'].drop_duplicates().sort_values().drop_duplicates()
gold_drug = drug_indi['DrugBank_ID'].drop_duplicates()
indep_drug = indep_indi['DrugBank_ID'].drop_duplicates()

binary_drug = pd.DataFrame(columns=gold_drug.tolist(), index=umls_ids.tolist())
binary_drug = binary_drug.fillna(0)
print len(binary_drug)
binary_drug.head(1)

f = lambda x: 1 if x else 0
    
for dg_id in gold_drug:
    dg_df = drug_indi[drug_indi.DrugBank_ID == dg_id]
    umls_id = dg_df['UMLS_concept_id_for_MedDRA']
    flags = umls_ids.isin(umls_id)
    oz_flags = map(f, flags)
    binary_drug[dg_id] = oz_flags

binary_drug.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/sider_indication/3_4_gold_indi_matrix.tsv", sep='\t')
binary_drug.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/sider_indication/3_4_gold_indi_matrix.tsv", sep='\t')

indep -indication matrix 만들기

binary_indep = pd.DataFrame(columns=indep_drug.tolist(), index = umls_ids.tolist())
binary_indep = binary_indep.fillna(0)
print len(binary_indep)
binary_indep.head(1)

f = lambda x: 1 if x else 0
    
for dg_id in indep_drug:
    dg_df = indep_indi[indep_indi.DrugBank_ID == dg_id]
    umls_id = dg_df['UMLS_concept_id_for_MedDRA']
    flags = umls_ids.isin(umls_id)
    oz_flags = map(f, flags)
    binary_indep[dg_id] = oz_flags

binary_indep.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/sider_indication/4_2_indep_indi_matrix.tsv", sep='\t')
binary_indep.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/sider_indication/4_2_indep_indi_matrix.tsv", sep='\t')
