==> Drug_Repositioning/8_new_training/PREDICT_data/3_descriptor_origianl_and_indication.ipynb <==
import pandas as pd
import numpy as np
from itertools import product
from datetime import datetime

#gold_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/2_2_gold_learning_asso.tsv")
#indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/3_2_indep_learning_asso.tsv")
gold_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/6_3_gold_learning_asso_indi.tsv")
indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/7_2_indep_learning_asso_indi.tsv")
drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/4_2_gold_se_by_indep_matrix.tsv", index_col = 0)
disease = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col = 0)
indi = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/sider_indication/3_4_gold_indi_matrix.tsv", index_col=0)

print len(gold_asso)
gold_asso.head(1)

print len(drug)
drug.head(1)

print len(disease)
disease.head(1)

print len(indi)
indi.head(1)

start = 0
end = 10001
cnt = start
arr = np.empty((0,5883), dtype=np.int64)
arr = np.hstack((['drug_id', 'disease_id', 'association', drug.index, disease.index, indi.index]))
print "start : %d : " % cnt + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
for i in gold_asso.values[start:end]:
    association = i[0]
    disease_id = i[1]
    drug_id = i[2]
    feature = np.hstack([drug_id, disease_id, association, drug[drug_id], disease[str(disease_id)], indi[drug_id]])
    arr = np.vstack([arr, feature])
    cnt = cnt + 1
    if cnt % 1000 == 0:
        print "%d : " % cnt + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print "end : %d : " % cnt + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print cnt
print arr.shape
df_arr = pd.DataFrame(data = arr[1:],columns = arr[0])
df_arr.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/3_1_gold_indi_descriptor_{}_{}tsv".format(start, end-1), sep='\t', index = False)

df_arr.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/2_1_gold_descriptor_{}_{}.tsv".format(start, end-1), sep='\t', index = False)

path = '/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/1_feature_based-Systematic/3_indication/'
arr1 = pd.read_table(path + "5_1_gold_indi_descriptor_0_10000tsv")
arr2 = pd.read_table(path + "5_1_gold_indi_descriptor_10001_20000tsv")
arr3 = pd.read_table(path + "5_1_gold_indi_descriptor_20001_30000tsv")
arr4 = pd.read_table(path + "5_1_gold_indi_descriptor_30001_40000tsv")
arr5 = pd.read_table(path + "5_1_gold_indi_descriptor_40001_50000tsv")
arr6 = pd.read_table(path + "5_1_gold_indi_descriptor_50001_60000tsv")
arr7 = pd.read_table(path + "5_1_gold_indi_descriptor_60001_70000tsv")
arr8 = pd.read_table(path + "5_1_gold_indi_descriptor_70001_88884tsv")
#arr9 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/2_1_gold_descriptor_80001_90000.tsv")
#arr10 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/2_1_gold_descriptor_90001_100000.tsv")
#arr11 = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/2_1_gold_descriptor_100001_103145.tsv")

frames = [arr1, arr2, arr3, arr4, arr5, arr6, arr7, arr8] #, arr9, arr10, arr11]
df_arr_all = pd.concat(frames , axis=0)
print len(df_arr_all)
df_arr_all.head(1)

df_arr_all.to_csv(path + "5_1_gold_indi_descriptor_combined.tsv", sep = '\t', index = False)

indep_drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/3_2_indep_drug_se_matrix.tsv", index_col = 0)
indep_indi = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/sider_indication/4_2_indep_indi_matrix.tsv", index_col = 0)

#independent
start = 0
end = len(indep_asso)
cnt = start
indep_arr = np.empty((0,5883), dtype=np.int64)
indep_arr = np.hstack((['drug_id', 'disease_id', 'association', drug.index, disease.index, indep_indi.index]))
print end
print "start : %d : " % cnt + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
for i in indep_asso.values[start:end]:
    association = i[2]
    disease_id = i[1]
    drug_id = i[0]
    feature = np.hstack([drug_id, disease_id, association, indep_drug[drug_id], disease[str(disease_id)], indep_indi[drug_id]])
    indep_arr = np.vstack([indep_arr, feature])
    cnt = cnt + 1
    if cnt % 1000 == 0:
        print "%d : " % cnt + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print "end : %d : " % cnt + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print cnt
print indep_arr.shape
df_indep_arr = pd.DataFrame(data = indep_arr[1:],columns = indep_arr[0])
df_indep_arr.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/5_2_indep_indi_descriptor.tsv", sep='\t', index = False)

indi = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/sider_indication/3_4_gold_indi_matrix.tsv", index_col=0)

5_1_gold_indi_descriptor_combined

df = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/5_1_gold_indi_descriptor_combined.tsv")

df.columns.values

print df.columns.values[3:1848]
print df.columns.values[1848:5370]
print df.columns.values[1848:5368]
print df.columns.values[5369:]

# indication descriptor의 위치를 바꾼다
gold = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/5_1_gold_indi_descriptor_combined.tsv")
indep = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/5_2_indep_indi_descriptor.tsv")

gold.head(1)

indep.head(1)

print gold.columns.values[3:1848]
print gold.columns.values[1848:5370]
print gold.columns.values[1848:5368]
print gold.columns.values[5369:]

print gold.columns.values[3:1847] # drug
print gold.columns.values[1848:] # disease
print gold.columns.values[3:]
