==> Drug_Repositioning/8_new_training/PREDICT_data/5_gold_n_indep_p_distance.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime
from sklearn.metrics import jaccard_similarity_score
from __future__ import division
from itertools import product

gold = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/4_1_tn_whole_descriptor_bayesian_predict.tsv")
indep = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/PREDICT_data/2_2_indep_descriptor.tsv")

gold_p = gold[gold.association == 1]
gold_n = gold[gold.association == 0]
indep_p = indep[indep.association == 1]
indep_n = indep[indep.association == 0]
print len(gold), len(gold_p), len(gold_n)
print len(indep), len(indep_p), len(indep_n)

gold_p.head(1)

gold_p_pair = gold_p.values[:, :2].tolist()
gold_n_pair = gold_n.values[:, :2].tolist()
indep_p_pair = indep_p.values[:, :2].tolist()
indep_n_pair = indep_n.values[:, :2].tolist()

# distance between gold_p and indep_p
df_gp_ip = pd.DataFrame()
cnt = 0
print len(gold_p_pair)
print len(indep_p_pair)
print 'start : ' + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
for index, row in gold_p.iterrows():
    gold_dr_id = row['drug_id'] 
    gold_di_id = row['disease_id']
    desc1 = row[3:].tolist()
    for index2, row2 in indep_p.iterrows():
        indep_dr_id = row2['drug_id']
        indep_di_id = row2['disease_id']
        desc2 = row2[3:].tolist()
        
        sim = jaccard_similarity_score(desc1, desc2)
        
        #dataframe형
        temp_df = pd.DataFrame(data = [[gold_dr_id, gold_di_id, indep_dr_id, indep_di_id, sim]],
                               columns=['gold_drug', 'gold_disease', 'indep_drug', 'indep_disease', 'Similarity'])
        df_gp_ip = df_gp_ip.append(temp_df)
    cnt = cnt + 1
    if cnt % 1000 == 0:
        print 'cnt == {} : '.format(cnt) + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print 'end : ' + datetime.now().strftime('%Y-%m-%d %H:%M:%S')

# distance between gold_p and indep_n
df_gp_in = pd.DataFrame()
cnt = 0
print len(gold_p_pair)
print len(indep_n_pair)
print 'start : ' + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
for index, row in gold_p.iterrows():
    gold_dr_id = row['drug_id'] 
    gold_di_id = row['disease_id']
    desc1 = row[3:].tolist()
    for index2, row2 in indep_n.iterrows():
        indep_dr_id = row2['drug_id']
        indep_di_id = row2['disease_id']
        desc2 = row2[3:].tolist()
        
        sim = jaccard_similarity_score(desc1, desc2)
        
        #dataframe형
        temp_df = pd.DataFrame(data = [[gold_dr_id, gold_di_id, indep_dr_id, indep_di_id, sim]],
                               columns=['gold_drug', 'gold_disease', 'indep_drug', 'indep_disease', 'Similarity'])
        df_gp_ip = df_gp_ip.append(temp_df)
    cnt = cnt + 1
    if cnt % 1000 == 0:
        print 'cnt == {} : '.format(cnt) + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print 'end : ' + datetime.now().strftime('%Y-%m-%d %H:%M:%S')

# distance between gold_n and indep_p
df_gn_ip = pd.DataFrame()
print len(gold_n_pair)
print len(indep_p_pair)
cnt = 0
all_descs = list(product(gold_n.values[:, 3:],indep_p.values[:, 3:]))
all_ids = list(product(gold_n_pair,indep_p_pair))
print 'start : ' + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
for ids, descs in zip(all_ids, all_descs):
    gold_dr_id = ids[0][0]
    gold_di_id = ids[0][1]
    desc1 = descs[0].tolist()
    
    indep_dr_id = ids[1][0]
    indep_di_id = ids[1][1]
    desc2 = descs[1].tolist()
        
    sim = jaccard_similarity_score(desc1, desc2)
        
    #dataframe형
    temp_df = pd.DataFrame(data = [[gold_dr_id, gold_di_id, indep_dr_id, indep_di_id, sim]],
                               columns=['gold_drug', 'gold_disease', 'indep_drug', 'indep_disease', 'Similarity'])
    df_gn_ip = df_gn_ip.append(temp_df)
    cnt = cnt + 1
    if cnt % 10000 == 0:
        print 'cnt == {} : '.format(cnt) + datetime.now().strftime('%Y-%m-%d %H:%M:%S')
print 'end : ' + datetime.now().strftime('%Y-%m-%d %H:%M:%S')

# distance between gold_n and indep_n
